fix(symthaea): extract 50+ manager magic numbers to named constants
Drive manager (11): boredom increment/threshold/scale, flow arousal
window, flow LR/exploration/confidence/boredom-reset, surprise scales
(Berlyne 1960, Csikszentmihalyi 1990, Yerkes-Dodson 1908).

Perception manager (10): budget EMA/dampen, coherence confidence/
exploration scales, arousal correction gain, vigilance amplify/recovery,
binding confidence scale (Friston 2010, Treisman & Gelade 1980).

Learning manager (11): plasticity thresholds, LR floor/scale, dream
exploration/LR boost, error trend scales, dissipative health/somatic
stress sensitivity (BCM 1982, Walker & Stickgold 2006).

Memory phase (8): high-Phi promotion cadence, codebook diversity/
utilization intervals, utilization EMA, novelty threshold adaptation
rates and triggers (Dehaene 2014, Kohonen 1982).

+++ symthaea/src/cognitive_loop/thresholds.rs
@@ -3001,14 +3001,14 @@ pub const LEARNING_SOMATIC_STRESS_SENSITIVITY: f64 = 0.2;
 
 /// Co-prime cadence for high-Phi episode→codebook promotion.
 /// Basis: Dehaene (2014) — conscious access creates durable representations.
-pub const MEMORY_HIGH_PHI_PROMOTION_CADENCE: u64 = 97;
+pub const MEMORY_HIGH_PHI_PROMOTION_CADENCE: usize = 97;
 
 /// Codebook diversity computation amortization interval (cycles).
 /// Basis: Kohonen (1982) — periodic reorganization of self-organizing maps.
-pub const MEMORY_CODEBOOK_DIVERSITY_INTERVAL: u64 = 50;
+pub const MEMORY_CODEBOOK_DIVERSITY_INTERVAL: usize = 50;
 
 /// Codebook utilization rate update interval (cycles).
-pub const MEMORY_CODEBOOK_UTILIZATION_INTERVAL: u64 = 50;
+pub const MEMORY_CODEBOOK_UTILIZATION_INTERVAL: usize = 50;
 
 /// Codebook utilization EMA decay.
 pub const MEMORY_CODEBOOK_UTIL_EMA_DECAY: f32 = 0.8;
